Move pagesディレクトリ内にあった__tests__をルートディレクトリへ
